+++ test/bulk-edit/e2e.spec.ts
@@ -3,8 +3,8 @@ import type { PayloadTestSDK } from 'helpers/sdk/index.js'
 
 import { expect, test } from '@playwright/test'
 import { addArrayRow } from 'helpers/e2e/fields/array/index.js'
+import { addListFilter } from 'helpers/e2e/filters/index.js'
 import { selectInput } from 'helpers/e2e/selectInput.js'
-import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
 import { toggleBlockOrArrayRow } from 'helpers/e2e/toggleCollapsible.js'
 import * as path from 'path'
 import { wait } from 'payload/shared'
