+++ test/fields/collections/Checkbox/e2e.spec.ts
@@ -1,9 +1,8 @@
 import type { Page } from '@playwright/test'
 
 import { expect, test } from '@playwright/test'
-import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
+import { addListFilter } from 'helpers/e2e/filters/index.js'
 import path from 'path'
-import { wait } from 'payload/shared'
 import { fileURLToPath } from 'url'
 
 import { ensureCompilationIsDone, initPageConsoleErrorCatch } from '../../../helpers.js'
