+++ test/fields/collections/Email/e2e.spec.ts
@@ -1,10 +1,7 @@
 import type { Page } from '@playwright/test'
 
 import { expect, test } from '@playwright/test'
-import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
-import { openListFilters } from 'helpers/e2e/filters/openListFilters.js'
 import path from 'path'
-import { wait } from 'payload/shared'
 import { fileURLToPath } from 'url'
 
 import type { PayloadTestSDK } from '../../../helpers/sdk/index.js'
