+++ test/fields/collections/Text/e2e.spec.ts
@@ -2,10 +2,9 @@ import type { Page } from '@playwright/test'
 import type { GeneratedTypes } from 'helpers/sdk/types.js'
 
 import { expect, test } from '@playwright/test'
-import { openListColumns } from 'helpers/e2e/columns/openListColumns.js'
+import { openListColumns, toggleColumn } from 'helpers/e2e/columns/index.js'
+import { addListFilter } from 'helpers/e2e/filters/index.js'
 import { upsertPreferences } from 'helpers/e2e/preferences.js'
-import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
-import { toggleColumn } from 'helpers/e2e/columns/toggleColumn.js'
 import path from 'path'
 import { wait } from 'payload/shared'
 import { fileURLToPath } from 'url'
