+++ test/group-by/e2e.spec.ts
@@ -3,9 +3,8 @@ import type { PayloadTestSDK } from 'helpers/sdk/index.js'
 
 import { expect, test } from '@playwright/test'
 import { devUser } from 'credentials.js'
-import { sortColumn } from 'helpers/e2e/columns/sortColumn.js'
-import { toggleColumn } from 'helpers/e2e/columns/toggleColumn.js'
-import { addListFilter } from 'helpers/e2e/filters/addListFilter.js'
+import { sortColumn, toggleColumn } from 'helpers/e2e/columns/index.js'
+import { addListFilter } from 'helpers/e2e/filters/index.js'
 import { goToNextPage } from 'helpers/e2e/goToNextPage.js'
 import { addGroupBy, clearGroupBy, closeGroupBy, openGroupBy } from 'helpers/e2e/groupBy/index.js'
 import { deletePreferences } from 'helpers/e2e/preferences.js'
