+++ test/helpers/e2e/columns/toggleColumn.ts
@@ -4,6 +4,7 @@ import { expect } from '@playwright/test'
 
 import { exactText } from '../../../helpers.js'
 import { openListColumns } from './openListColumns.js'
+import { waitForColumnInURL } from './waitForColumnsInURL.js'
 
 export const toggleColumn = async (
   page: Page,
@@ -63,25 +64,4 @@ export const toggleColumn = async (
   }
 
   return { columnContainer }
-}
-
-export const waitForColumnInURL = async ({
-  page,
-  columnName,
-  state,
-}: {
-  columnName: string
-  page: Page
-  state: 'off' | 'on'
-}): Promise<void> => {
-  await page.waitForURL(/.*\?.*/)
-
-  const identifier = `${state === 'off' ? '-' : ''}${columnName}`
-
-  // Test that the identifier is in the URL
-  // It must appear in the `columns` query parameter, i.e. after `columns=...` and before the next `&`
-  // It must also appear in it entirety to prevent partially matching other values, i.e. between quotation marks
-  const regex = new RegExp(`columns=([^&]*${encodeURIComponent(`"${identifier}"`)}[^&]*)`)
-
-  await page.waitForURL(regex)
 }
