+++ test/query-presets/e2e.spec.ts
@@ -2,8 +2,7 @@ import type { BrowserContext, Page } from '@playwright/test'
 
 import { expect, test } from '@playwright/test'
 import { devUser } from 'credentials.js'
-import { openListColumns } from 'helpers/e2e/columns/openListColumns.js'
-import { toggleColumn } from 'helpers/e2e/columns/toggleColumn.js'
+import { openListColumns, toggleColumn } from 'helpers/e2e/columns/index.js'
 import { openNav } from 'helpers/e2e/toggleNav.js'
 import * as path from 'path'
 import { fileURLToPath } from 'url'
